myproject-server/pyproject.toml
[project]
name = "myproject-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nguyentran0212", email = "gen.tran.1991@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.128.0",
    "myproject-core",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.33",
]

[build-system]
requires = ["uv_build>=0.9.27,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
myproject-core = { workspace = true }
